backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
openai>=1.0.0
python-multipart>=0.0.6
pillow>=10.0.0
sqlalchemy>=2.0.21
pydantic>=2.4.2
pytest>=7.4.2
python-dotenv>=1.0.0
numpy>=1.25.2
matplotlib>=3.8.0
opencv-python>=4.8.0.76
langchain>=0.0.300
